--- frontend/helpers/getTime.tsx
@@ -1,9 +1,0 @@
-export default function getTime (time: Date) {
-  const hours = time.getHours();
-  const minutes = time.getMinutes();
-  var sHours = hours.toString();
-  var sMinutes = minutes.toString();
-  if(hours<10) sHours = "0" + sHours;
-  if(minutes<10) sMinutes = "0" + sMinutes;
-  return `${sHours}:${sMinutes}`
-}
