+++ frontend/helpers/index.tsx
@@ -1,3 +1,5 @@
 export * from '@/helpers/a11yProps';
 export * from '@/helpers/filterHelpers';
 export * from '@/helpers/finderHelpers';
+export * from '@/helpers/getHelpers';
+export * from '@/helpers/parsingHelpers';
