--- frontend/helpers/returnFirstKey.tsx
@@ -1,4 +1,0 @@
-export default function returnFirstKey (object: any) {
-  var keys = Object.keys(object);
-  return keys[0]
-}
