--- frontend/helpers/trueOrFalse.tsx
@@ -1,7 +1,0 @@
-export default function trueOrFalse (value: 'Yes' | 'No') {
-  if (value === 'Yes') {
-    return true
-  }
-
-  return false
-}
