+++ frontend/pages/_app.tsx
@@ -1,18 +1,16 @@
 'use client'
 
-import { ReactElement, ReactNode, useEffect, useState } from 'react'
-import type { NextPage } from 'next'
-import type { AppProps } from 'next/app'
-import Head from 'next/head'
-import { Analytics } from '@vercel/analytics/react'
+import { ReactElement, ReactNode, useEffect, useState } from 'react';
+import type { NextPage } from 'next';
+import type { AppProps } from 'next/app';
+import Head from 'next/head';
+import { Analytics } from '@vercel/analytics/react';
 
-import parseData from "@/helpers/parseData"
+import '@/assets/styles/application.scss';
+import { Navigation } from '@/components';
+import { parseData } from "@/helpers";
 
-export const { eventData, filteredEvents } = parseData()
-
-import '@/assets/styles/application.scss'
-
-import { Navigation } from '@/components'
+export const { eventData, filteredEvents } = parseData();
 
 export const metadata = {
   title: 'GenCon Cal',
